auth_services/requirements.txt
# Django Framework
Django==5.1.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7

# Authentication & Security
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
python-dotenv==1.0.1
social-auth-app-django==5.6.0

# Database
psycopg2-binary==2.9.9

# Google OAuth
google-auth
google-auth-oauthlib
google-api-python-client

# MFA, OTP, QR Codes
pyotp==2.9.0
qrcode==7.4.2

# Utilities
requests==2.31.0

# Deployment
gunicorn==21.2.0

#rabbitMQ
pika==1.3.2
djangorestframework-simplejwt

Pillow

# Celery
celery==5.3.6
redis==5.0.1
cloudinary>=1.36.0

